Disable fix, language menu, lang order.
Fix:  A disabled mod could prevent language packs
from loading.

Added a "Set Language" menu item at the main menu
to make it easier to switch to the language.

Moved new languages to the top of the languages List
so it is in view first.

Moved patches to new location.

Updated TestLanguage to have back and quit both translated
for immediate feedback when the user changes their language.